+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com), and this project adheres to [Semantic Versioning](https://semver.org).
 
+## [0.9.0] - 2021-12-04
+
+### Added
+
+- Button component with a built-in icon.
+
 ## [0.8.0] - 2021-12-04
 
 ### Added
@@ -84,7 +90,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com), and this 
 - Card component.
 - Material icon component.
 
-[unreleased]: https://github.com/visiosto/react-components/compare/v0.8.0...HEAD
+[unreleased]: https://github.com/visiosto/react-components/compare/v0.9.0...HEAD
+[0.9.0]: https://github.com/visiosto/react-components/compare/v0.8.0...v0.9.0
 [0.8.0]: https://github.com/visiosto/react-components/compare/v0.7.0...v0.8.0
 [0.7.0]: https://github.com/visiosto/react-components/compare/v0.6.0...v0.7.0
 [0.6.0]: https://github.com/visiosto/react-components/compare/v0.5.0...v0.6.0